+++ src/main/java/xyz/theprogramsrc/supercoreapi/Base.java
@@ -1,7 +1,7 @@
 package xyz.theprogramsrc.supercoreapi;
 
 
-import xyz.theprogramsrc.supercoreapi.global.updater.SpigotUpdateChecker;
+import xyz.theprogramsrc.supercoreapi.global.updater.SongodaUpdateChecker;
 
 /**
  * This is a set of utils that need to be loaded
@@ -16,17 +16,16 @@ public class Base {
     }
 
     private void loadUpdateChecker(){
-        new SpigotUpdateChecker("77096"){
+        new SongodaUpdateChecker("supercoreapi-the-best-way-to-create-a-plugin"){
             @Override
-            public void onCheckFail() {
+            public void onFailCheck() {
                 Base.this.plugin.log("&cError while checking &4SuperCoreAPI &cupdates.");
             }
 
             @Override
-            public void onCheckSuccess(String lastVersion) {
-                String currentVersion = "4.2.9";
-                int latest = Integer.parseInt(lastVersion.split(" ")[0].replaceAll("\\.", ""));
-                int current = Integer.parseInt(currentVersion.split(" ")[0].replaceAll("\\.", ""));
+            public void onSuccessCheck(String lastVersion) {
+                int latest = Integer.parseInt(lastVersion.split(" ")[0].replaceAll("\\.", "").replaceAll("[a-zA-Z]", ""));
+                int current = Integer.parseInt(Base.this.plugin.getPluginVersion().split(" ")[0].replaceAll("\\.", "").replaceAll("[a-zA-Z]", ""));
                 if(latest > current){
                     Base.this.plugin.log("&cNew update available for &3SuperCoreAPI &7("+lastVersion+")");
                     Base.this.plugin.log("&cIf you're the developer of the plugin update the API, if you're a customer of a plugin using &3SuperCoreAPI&c please notify to the Developer.");
@@ -35,6 +34,6 @@ public class Base {
                     Base.this.plugin.log("&cUnless you're the developer, you must use the version &7" + lastVersion);
                 }
             }
-        }.check();
+        }.checkUpdates();
     }
 }