+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## v4.2.12 Changelog:
+```
+* Migrated Update Checker to Songoda
+* Bug Fixes
+```
+
+## v4.2.11 Changelog:
+```
+- Removed tests in SpigotPlugin
+```
+
 ## v4.2.10 Changelog:
 ```
 + Added Scroll Browser (Vertical Browser)